Overhaul w/ AI help 🐳
This has been streamlined tremendously w/ Deepseek help. It's been changed to use fuzzy finder as a way to find all the projects I have currently in my project directory as well as have an option to just go to the root of my project folder.

It now depends on fuzzy finder as a dependecy which is fine